backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
psycopg2-binary
sqlalchemy
alembic
bcrypt
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
psycopg2